authentication implemented with jwt token. posts can be edited or deleted only by creators

+++ controllers/feed.js
@@ -3,13 +3,30 @@ const path = require('path');
 const fs = require('fs');
 
 const Post = require(path.join(__dirname, '../models/post'));
+const User = require(path.join(__dirname, '../models/user'));
 
 exports.getPosts = (req, res, next) => {
-    Post.find()
+    const currentPage = req.query.page || 1;
+    const perPage = 2;
+    let totalItems;
+    Post
+        .find()
+        .countDocuments()
+        .then(count => {
+            totalItems = count;
+            return Post
+                .find()
+                .skip((currentPage - 1) * perPage)
+                .limit(perPage);
+        })
         .then(posts => {
             res
                 .status(200)
-                .json({ message: 'Fetched posts successfully.', posts: posts });
+                .json({
+                    message: 'Fetched posts successfully.',
+                    posts: posts,
+                    totalItems: totalItems
+                });
         })
         .catch(err => {
             if (!err.statusCode) {
@@ -34,18 +51,28 @@ exports.createPost = (req, res, next) => {
     const imageUrl = req.file.path;
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
     const post = new Post({
         title: title,
         content: content,
         imageUrl: imageUrl,
-        creator: { name: 'Nedim' }
+        creator: req.userId
     });
     post
         .save()
         .then(result => {
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            creator = user;
+            user.posts.push(post);
+            return user.save();
+        })
+        .then(result => {
             res.status(201).json({
                 message: 'Post created successfully!',
-                post: result
+                post: post,
+                creator: {_id: creator._id, name: creator.name}
             });
         })
         .catch(err => {
@@ -65,7 +92,7 @@ exports.getPost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
-            res.status(200).json({ message: 'Post fetched.', post: post });
+            res.status(200).json({message: 'Post fetched.', post: post});
         })
         .catch(err => {
             if (!err.statusCode) {
@@ -101,6 +128,11 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized.');
+                error.statusCode = 403;
+                throw error;
+            }
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl);
             }
@@ -110,7 +142,7 @@ exports.updatePost = (req, res, next) => {
             return post.save();
         })
         .then(result => {
-            res.status(200).json({ message: 'Post updated!', post: result });
+            res.status(200).json({message: 'Post updated!', post: result});
         })
         .catch(err => {
             if (!err.statusCode) {
@@ -130,11 +162,22 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized.');
+                error.statusCode = 403;
+                throw error;
+            }
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
         })
         .then(result => {
-            console.log(result);
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
+        })
+        .then(result => {
             res.status(200).json({
                 message: 'Post deleted.'
             });
